adjust some tests

+++ src/tests/partial_order_fill.rs
@@ -6,10 +6,10 @@ use crate::{mock_exchange_linear, prelude::*, trade, DECIMALS};
 #[test_case(QuoteCurrency::new(100, 0), BaseCurrency::new(2, 0), Side::Buy, QuoteCurrency::new(99, 0); "With buy order")]
 #[test_case(QuoteCurrency::new(101, 0), BaseCurrency::new(2, 0), Side::Sell, QuoteCurrency::new(102, 0); "With sell order")]
 fn partial_limit_order_fill(
-    limit_price: QuoteCurrency<i32, DECIMALS>,
-    qty: BaseCurrency<i32, DECIMALS>,
+    limit_price: QuoteCurrency<i64, DECIMALS>,
+    qty: BaseCurrency<i64, DECIMALS>,
     side: Side,
-    trade_price: QuoteCurrency<i32, DECIMALS>,
+    trade_price: QuoteCurrency<i64, DECIMALS>,
 ) {
     let mut exchange = mock_exchange_linear();
